fix: move automatic json parser management to json viz scene
